backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0
aiosmtplib>=3.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.1
slowapi>=0.1.9
email-validator>=2.1.1
pydantic[email]>=2.6.0
PyJWT>=2.8.0
python-multipart>=0.0.9
aiofiles>=23.2.1
